Move `spinner` to its own file
