+++ lib/streamlit/__init__.py
@@ -54,10 +54,8 @@ _LOGGER = _logger.get_logger("root")
 # Give the package a version.
 __version__ = _STREAMLIT_VERSION_STRING
 
-from typing import Any, Dict, Iterator, List, NoReturn
-import contextlib as _contextlib
+from typing import Any, Dict, List, NoReturn
 import sys as _sys
-import threading as _threading
 import urllib.parse as _parse
 
 import click as _click
@@ -91,6 +89,7 @@ from streamlit.runtime.caching import (
     singleton as experimental_singleton,
     memo as experimental_memo,
 )
+from streamlit.spinner import spinner as spinner
 
 cache = _gather_metrics(_cache)
 
@@ -388,67 +387,6 @@ def experimental_set_query_params(**query_params: Any) -> None:
     msg = _ForwardMsg_pb2.ForwardMsg()
     msg.page_info_changed.query_string = ctx.query_string
     ctx.enqueue(msg)
-
-
-@_contextlib.contextmanager
-def spinner(text: str = "In progress...") -> Iterator[None]:
-    """Temporarily displays a message while executing a block of code.
-
-    Parameters
-    ----------
-    text : str
-        A message to display while executing that block
-
-    Example
-    -------
-
-    >>> with st.spinner('Wait for it...'):
-    >>>     time.sleep(5)
-    >>> st.success('Done!')
-
-    """
-    import streamlit.runtime.legacy_caching.caching as legacy_caching
-    import streamlit.runtime.caching as caching
-    from streamlit.proto.Spinner_pb2 import Spinner as SpinnerProto
-    from streamlit.string_util import clean_text
-
-    # @st.cache optionally uses spinner for long-running computations.
-    # Normally, streamlit warns the user when they call st functions
-    # from within an @st.cache'd function. But we do *not* want to show
-    # these warnings for spinner's message, so we create and mutate this
-    # message delta within the "suppress_cached_st_function_warning"
-    # context.
-    with legacy_caching.suppress_cached_st_function_warning():
-        with caching.suppress_cached_st_function_warning():
-            message = empty()
-
-    try:
-        # Set the message 0.1 seconds in the future to avoid annoying
-        # flickering if this spinner runs too quickly.
-        DELAY_SECS = 0.1
-        display_message = True
-        display_message_lock = _threading.Lock()
-
-        def set_message():
-            with display_message_lock:
-                if display_message:
-                    with legacy_caching.suppress_cached_st_function_warning():
-                        with caching.suppress_cached_st_function_warning():
-                            spinner_proto = SpinnerProto()
-                            spinner_proto.text = clean_text(text)
-                            message._enqueue("spinner", spinner_proto)
-
-        _add_script_run_ctx(_threading.Timer(DELAY_SECS, set_message)).start()
-
-        # Yield control back to the context.
-        yield
-    finally:
-        if display_message_lock:
-            with display_message_lock:
-                display_message = False
-        with legacy_caching.suppress_cached_st_function_warning():
-            with caching.suppress_cached_st_function_warning():
-                message.empty()
 
 
 @_gather_metrics
